requirements.txt
python-docx==1.1.2
streamlit==1.31.0
pandas==2.2.0
openpyxl==3.1.2
openai>=1.12.0
httpx>=0.24.0
python-dotenv==1.0.0
